+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - **feat**: date/addUTC now accepts both a date and a date string
 - **feat**: date/diff now accepts both a date and a date string
 - **feat**: date/endOfUTC now accepts both a date and a date string
+- **feat**: date/isLeap now accepts both a date and a date string
 - **feat**: date/format now accepts both a date and a date string
 - **feat**: date/setTimeUTC now accepts both a date and a date string
 - **feat**: date/startOfUTC now accepts both a date and a date string